Export estimator to the separate file
